+++ scripts/reading.js
@@ -127,7 +127,7 @@ function goToImageCL(index, animation = true, fromScroll = false, fromPageRange 
 		music.focusIndex(index);
 	}
 
-	let animationDurationMS = ((animation) ? config.readingViewSpeed : 0) * 1000;
+	let animationDurationMS = ((animation) ? _config.readingViewSpeed : 0) * 1000;
 	let contentLeft = template._contentLeft();
 
 	let leftScroll = contentLeft.firstElementChild;
@@ -154,7 +154,7 @@ function goToImageCL(index, animation = true, fromScroll = false, fromPageRange 
 		if(leftItem)
 		{
 			leftItem.classList.add('s');
-			if(animation && config.readingViewSpeed > 0.2) leftItem.classList.add('transition');
+			if(animation && _config.readingViewSpeed > 0.2) leftItem.classList.add('transition');
 		}
 	}
 	else
@@ -169,7 +169,7 @@ function goToImageCL(index, animation = true, fromScroll = false, fromPageRange 
 		if(leftItem)
 		{
 			leftItem.classList.add('s');
-			if(animation && config.readingViewSpeed > 0.2) leftItem.classList.add('transition');
+			if(animation && _config.readingViewSpeed > 0.2) leftItem.classList.add('transition');
 		}
 	}
 
@@ -384,7 +384,7 @@ function goScrollPercent(screenPercent = 50, animation = true)
 {
 	if(_config.readingWebtoon || readingViewIs('scroll'))
 	{
-		let animationDurationMS = (animation ? config.readingViewSpeed : 0) * 1000;
+		let animationDurationMS = (animation ? _config.readingViewSpeed : 0) * 1000;
 
 		const content = template._contentRight().firstElementChild;
 		const rect = content.getBoundingClientRect();
@@ -418,7 +418,7 @@ function goScrollPercent(screenPercent = 50, animation = true)
 //Go to a specific comic index
 function goToIndex(index, animation = true, nextPrevious = false, end = false)
 {
-	let animationDurationS = ((animation) ? config.readingViewSpeed : 0);
+	let animationDurationS = ((animation) ? _config.readingViewSpeed : 0);
 	let animationDurationMS = animationDurationS * 1000;
 
 	let _currentScale = currentScale;
@@ -751,13 +751,13 @@ function goPrevComic()
 	
 	if(hasComic)
 	{
-		let speed = config.readingViewSpeed;
-		config.readingViewSpeed = 0;
+		let speed = _config.readingViewSpeed;
+		_config.readingViewSpeed = 0;
 
 		let double = goStart();
 		if(double) goStart();
 
-		config.readingViewSpeed = speed;
+		_config.readingViewSpeed = speed;
 	}
 }
 
@@ -797,13 +797,13 @@ function goNextComic()
 
 	if(hasComic)
 	{
-		let speed = config.readingViewSpeed;
-		config.readingViewSpeed = 0;
+		let speed = _config.readingViewSpeed;
+		_config.readingViewSpeed = 0;
 
 		let double = goEnd();
 		if(double) goEnd();
 
-		config.readingViewSpeed = speed;
+		_config.readingViewSpeed = speed;
 	}
 }
 
@@ -1031,7 +1031,7 @@ function showNextComic(mode, animation = true, invert = false)
 
 	if(mode == 1)
 	{
-		var transition = config.readingViewSpeed < config.readingDelayComicSkip ? config.readingViewSpeed : config.readingDelayComicSkip;
+		var transition = _config.readingViewSpeed < config.readingDelayComicSkip ? _config.readingViewSpeed : config.readingDelayComicSkip;
 
 		if(config.readingDelayComicSkip != 0)
 		{
@@ -1089,31 +1089,31 @@ function showNextComic(mode, animation = true, invert = false)
 		if(readingViewIs('scroll'))
 		{
 			var skip = template.contentRight('.reading-skip-bottom').css({
-				'transition': config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': _config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'translate(0px, 0px)',
 				'opacity': 0,
 			});
 
 			template.contentRight('.reading-body').css({
-				'transition': 'transform '+config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': 'transform '+_config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'scale(1) translate(0px, 0px)',
 			});
 		}
 		else
 		{
 			var skip = template.contentRight('.reading-skip-right').css({
-				'transition': config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': _config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'translate(0px, 0px)',
 				'opacity': 0,
 			});
 
 			template.contentRight('.reading-body > div').css({
-				'transition': 'transform '+config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': 'transform '+_config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'scale(1) translate(-'+(contentWidth * (readingViewIs('slide') ? (view.distribution.total - 1) : 0))+'px, 0px)',
 			});
 		}
 
-		showComicSkip = setTimeout(function(){showComicSkip = false}, config.readingViewSpeed * 1000);
+		showComicSkip = setTimeout(function(){showComicSkip = false}, _config.readingViewSpeed * 1000);
 		currentIndex = view.distribution.total;
 	}
 }
@@ -1129,7 +1129,7 @@ function showPreviousComic(mode, animation = true, invert = false)
 
 	if(mode == 1)
 	{
-		var transition = config.readingViewSpeed < config.readingDelayComicSkip ? config.readingViewSpeed : config.readingDelayComicSkip;
+		var transition = _config.readingViewSpeed < config.readingDelayComicSkip ? _config.readingViewSpeed : config.readingDelayComicSkip;
 
 		if(config.readingDelayComicSkip != 0)
 		{
@@ -1189,13 +1189,13 @@ function showPreviousComic(mode, animation = true, invert = false)
 			var skip = template.contentRight('.reading-skip-top');
 
 			skip.css({
-				'transition': config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': _config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'translate(0px, 0px)',
 				'opacity': 0,
 			});
 
 			template.contentRight('.reading-body').css({
-				'transition': 'transform '+config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': 'transform '+_config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'scale(1) translate(0px, 0px)',
 			});
 		}
@@ -1204,18 +1204,18 @@ function showPreviousComic(mode, animation = true, invert = false)
 			var skip = template.contentRight('.reading-skip-left');
 
 			skip.css({
-				'transition': config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': _config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'translate(0px, 0px)',
 				'opacity': 0,
 			});
 
 			template.contentRight('.reading-body').css({
-				'transition': 'transform '+config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
+				'transition': 'transform '+_config.readingViewSpeed+'s, background-color 0.2s, box-shadow 0.2s',
 				'transform': 'scale(1) translate(0px, 0px)',
 			});
 		}
 
-		showComicSkip = setTimeout(function(){showComicSkip = false}, config.readingViewSpeed * 1000);
+		showComicSkip = setTimeout(function(){showComicSkip = false}, _config.readingViewSpeed * 1000);
 		currentIndex = 1;
 	}
 }
@@ -1227,7 +1227,7 @@ function applyScale(animation = true, scale = 1, center = false, zoomOut = false
 	if((!onReading && !_onReading) || !isLoaded)
 		return;
 
-	let animationDurationS = ((animation) ? config.readingViewSpeed : 0);
+	let animationDurationS = ((animation) ? _config.readingViewSpeed : 0);
 
 	if(currentZoomIndex === false)
 	{
@@ -1449,7 +1449,7 @@ function applyScale(animation = true, scale = 1, center = false, zoomOut = false
 						zIndex: scale == 1 ? 1 : 3,
 					});
 
-				}, config.readingViewSpeed * 1000);
+				}, _config.readingViewSpeed * 1000);
 			}
 			else
 			{
@@ -1868,7 +1868,7 @@ function dragZoom(dx, dy, animation = false)
 
 function applyZoom(x, y, animation)
 {
-	const transition = app.scrollTransition('dragZoom', animation ? config.readingViewSpeed : 0);
+	const transition = app.scrollTransition('dragZoom', animation ? _config.readingViewSpeed : 0);
 	const withLimits = notCrossZoomLimits(x, y);
 
 	const diff = {
@@ -2516,6 +2516,14 @@ function changePagesView(mode, value, save)
 	{
 		updateReadingPagesConfig('readingViewAdjustToWidth', value);
 		reading.reloadAnimated(false, imageIndex);
+	}
+	else if(mode == 4) // Set the speed of the animation when changing pages
+	{
+		if(save) updateReadingPagesConfig('readingViewSpeed', value);
+	}
+	else if(mode == 5) // Set the delay when skip from comic
+	{
+		if(save) updateReadingPagesConfig('readingDelayComicSkip', value);
 	}
 	else if(mode == 6) // Set the reading to double page
 	{